# fairy/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "swagger接口文档——菜单权限管理",
        "title": "Swagger权限管理",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5001",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "用户登录系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sys-系统相关"
                ],
                "summary": "用户登录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "注册用户数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sys-系统相关"
                ],
                "summary": "注册用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}